requirements.txt
alembic==1.7.4
asgiref==3.4.1
asn1crypto==1.4.0
click==8.0.3
Cython==0.29.24
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.3
openpyxl==3.0.9
p5py==1.0.0
pep517==0.12.0
pg8000==1.22.0
python-dotenv==0.19.1
redis==3.5.3
scramp==1.4.1
SQLAlchemy==1.4.26
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
WTForms==2.3.3
